requirements.txt
json5
mysql-connector-python
pandas
requests
flask
Flask-APScheduler
openpyxl
waitress
scipy
pyarrow